Add Trait for Singleton and Data objects
